Remove copy between tagger and dogstatsd metrics
When enriching tags for each dogstatsd point we do not need the copy.
The removed copy use to create a lot of allocation (1 per dogstatsd
point) and longer GC runs.

The copy was there to prevent modification of the dat in the cache from
outside the tagger. Since enrichTags only read the data it's safe.